fix: add example - Date serialization in responses
