+++ examples/7-responses-with-date-object/api.yaml
@@ -1,0 +1,50 @@
+openapi: '3.0.0'
+info:
+  version: 1.0.0
+  title: Swagger Petstore schemaObjectMapper variant
+  description: A sample API
+  termsOfService: http://swagger.io/terms/
+  license:
+    name: Apache 2.0
+    url: https://www.apache.org/licenses/LICENSE-2.0.html
+servers:
+  - url: /v1
+paths:
+  /date-time:
+    get:
+      responses:
+        200:
+          description: date-time handler
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  created_at:
+                    type: string
+                    format: date-time
+                  id:
+                    type: number
+  /date:
+    get:
+      responses:
+        200:
+          description: date handler
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/User'
+                  
+components:
+  schemas:
+    Date:
+      type: string
+      format: date
+    User:
+      type: object
+      properties:
+        id: 
+          type: number
+        created_at:
+          $ref: "#/components/schemas/Date"
+
