+++ examples/7-responses-with-date-object/package.json
@@ -12,16 +12,11 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "ajv": "^6.12.4",
     "body-parser": "^1.19.0",
-    "cookie-parser": "^1.4.5",
-    "express-openapi-validator": "^4.0.2",
-    "lodash": "^4.17.20",
-    "mongodb": "^3.6.0",
-    "morgan": "^1.10.0"
+    "express-openapi-validator": "file:../.."
   },
   "devDependencies": {
-    "nodemon": "^2.0.4",
+    "nodemon": "^2.0.6",
     "prettier": "^2.1.1"
   }
 }
