--- examples/7-schema-object-mapper/api.yaml
@@ -1,236 +1,0 @@
-openapi: '3.0.0'
-info:
-  version: 1.0.0
-  title: Swagger Petstore schemaObjectMapper variant
-  description: A sample API
-  termsOfService: http://swagger.io/terms/
-  license:
-    name: Apache 2.0
-    url: https://www.apache.org/licenses/LICENSE-2.0.html
-servers:
-  - url: /v1
-paths:
-  /ping:
-    get:
-      description: |
-        ping then pong!
-      operationId: ping
-      responses:
-        '200':
-          description: OK
-          content:
-            text/plain:
-              schema:
-                type: string
-                example: pong
-        default:
-          description: unexpected error
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Error'
-  /pets:
-    get:
-      description: |
-        Returns all pets
-      operationId: findPets
-      parameters:
-        - name: type
-          in: query
-          description: maximum number of results to return
-          required: true
-          schema:
-            type: string
-            enum:
-              - dog
-              - cat
-        - name: tags
-          in: query
-          description: tags to filter by
-          required: false
-          style: form
-          schema:
-            type: array
-            items:
-              type: string
-        - name: limit
-          in: query
-          description: maximum number of results to return
-          required: true
-          schema:
-            type: integer
-            format: int32
-            minimum: 1
-            maximum: 20
-      responses:
-        '200':
-          description: pet response
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  $ref: '#/components/schemas/Pet'
-        default:
-          description: unexpected error
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Error'
-
-    post:
-      description: Creates a new pet in the store.
-      operationId: addPet
-      security:
-        - ApiKeyAuth: []
-      requestBody:
-        description: Pet to add to the store
-        required: true
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/Pet'
-      responses:
-        '200':
-          description: pet response
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Pet'
-        default:
-          description: unexpected error
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Error'
-
-  /pets/{id}:
-    get:
-      description: Returns a user based on a single ID, if the user does not have access to the pet
-      operationId: find pet by id
-      parameters:
-        - name: id
-          in: path
-          description: ID of pet to fetch
-          required: true
-          schema:
-            $ref: '#/components/schemas/ObjectId'
-      responses:
-        '200':
-          description: pet response
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Pet'
-        default:
-          description: unexpected error
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Error'
-    delete:
-      description: deletes a single pet based on the ID supplied
-      operationId: deletePet
-      parameters:
-        - name: id
-          in: path
-          description: ID of pet to delete
-          required: true
-          schema:
-            $ref: '#/components/schemas/ObjectId'
-      responses:
-        '204':
-          description: pet deleted
-        default:
-          description: unexpected error
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Error'
-
-  /pets/{id}/photos:
-    post:
-      description: upload a photo of the pet
-      operationId: uploadPetPhoto
-      parameters:
-        - name: id
-          in: path
-          description: ID of pet to fetch
-          required: true
-          schema:
-            $ref: '#/components/schemas/ObjectId'
-      requestBody:
-        content:
-          multipart/form-data:
-            schema:
-              # $ref: '#/components/schemas/NewPhoto'
-              type: object
-              required:
-                - file
-              properties:
-                file:
-                  description: The photo
-                  type: string
-                  format: binary
-        required: true
-      responses:
-        201:
-          description: Created
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  success:
-                    type: boolean
-
-components:
-  schemas:
-    ObjectId:
-      type: string
-      pattern: '^[0-9a-fA-F]{24}$'
-    Date:
-      type: string
-      format: date
-    DateTime:
-      type: string
-      format: date-time
-    Pet:
-      required:
-        - id
-        - name
-        - type
-      properties:
-        id:
-          $ref: '#/components/schemas/ObjectId'
-        name:
-          type: string
-        tag:
-          type: string
-        type:
-          $ref: '#/components/schemas/PetType'
-        creationDate:
-          $ref: '#/components/schemas/DateTime'
-
-    PetType:
-      type: string
-      enum:
-        - dog
-        - cat
-
-    Error:
-      required:
-        - code
-        - message
-      properties:
-        code:
-          type: integer
-          format: int32
-        message:
-          type: string
-
-  securitySchemes:
-    ApiKeyAuth:
-      type: apiKey
-      in: header
-      name: X-API-Key
